\encoding{utf-8}
\name{deprecated}
\alias{deprecated}
\title{Deprecated functions in PowerTOST}
\description{
Details about deprecated functions in PowerTOST.
}
\details{
The functions \code{power.NTIDFDA}, \code{sampleN.NTIDFDA}, and \code{pa.NTIDFDA} were deprecated in version 1.5.4 and will be removed future versions.
}
\seealso{
\code{\link{power.NTID}}, \code{\link{sampleN.NTID}}, \code{\link{pa.NTIDFDA}} for the new functions.
}
