\name{type1error.2TOST}
\alias{type1error.2TOST}
\title{
Type I error rate for two simultaneous TOST procedures
}
\description{
Was designed to calculate the type I error rate of two simultaneous TOST procedures 
(where the two parameters of the two TOSTs are correlated with some correlation)
for various study designs used in BE studies.

Is defunct now since it suffers from insufficient precision to obtain the 
type 1 error (TIE) via simulations.\cr 
Due to the intersection-union principle the TIE is always upper bounded 
to alpha by theory.
}
