\name{PowerTOST-internal}
\title{Internal PowerTOST objects}
\alias{.approx.power.TOST}
\alias{.approx2.power.TOST}
\alias{.CV2se}
\alias{.design.bk}
\alias{.design.df}
\alias{.design.name}
\alias{.design.no}
\alias{.design.props}
\alias{.design.step}
\alias{.exppower.TOST}
\alias{.expsampleN0}
\alias{.power.TOST}
\alias{.Q.integrand}
\alias{.sampleN0}
\alias{.se2CV}
\description{Internal PowerTOST objects.}
\details{These are usually not to be called by the user.}
\keyword{internal}